Throwing Stools Works Again
Dropping a stool (not throwing) works the exact same as clicking it to
put it down (the result being a deconstructable 'structure' stool).

A thrown stool that isn't caught by a mob works the exact same as the
above (the result being a deconstructable 'structure' stool).

A thrown stool that IS caught by a mob remains in the throwable weapon
form.

+++ code/game/objects/structures/stool_bed_chair_nest/stools.dm
@@ -76,5 +76,14 @@
 
 /obj/item/weapon/stool/dropped(mob/user) //Make sure we aren't leaving un-deconstructible stools.
 	..()
-	origin.forceMove(user.loc) //Put the structure stool back down, not the weapon.
-	qdel(src)
+	if(!user.in_throw_mode) //If this stool isn't getting thrown, turn it back into a structure.
+		origin.forceMove(user.loc)
+		qdel(src)
+
+/obj/item/weapon/stool/throw_impact(atom/hit_atom) //Thrown but uncaught stools will automatically turn into the structure version as soon as the throw is over.
+	..()
+	if(!iscarbon(loc)) //If it wasn't caught by someone, put the structure version down.
+		var/atom/A = get_turf(src)
+		sleep(6) //Gets the work done at the last frame of the 'impact spin' animation so it doesn't look janky.
+		origin.forceMove(A)
+		qdel(src)
